logoff and new spinner
Because as now we can do only login, i added the logoff function to session.php after logging off
is sending back to login screen.
And added spinner on package search.

TODO:
get out function looks not working because adding it in restricting part of the website is giving
"headers already sent " error.