Removed var_dump on DBaction. Extended Varchar(20) to varchar(60) in DB creation. Added new function to DBAction. Fixed UpdateInstalledPackageList in DBAction. Fixed Search.php using DB instead of Restful. Added DebugFunction to Ajax and small js codefix. Added prototype php variable inject in js for Graph. DBcontroller Updating the packageList.
Because we decide to change back to DB instead of using Restful
directly for listing package. I'm reinstating the DBaction functions.
The search.php file called by ajax is now getting data from DB, for
implementing it, I need to remove var_dump on DbAction.php interfering
with the correct Ajax function and added a debug function for checking
the correct output from php.
Also added List package and search package function in DbAction.php
I also need to extend the DB Varchar in for the installed package and package
table for not cutting the name of the package, and making possible to check
for same package before updating.
DBcontroller is now updating the packageList still not doing async but need to
be called manually.
Added prototype for making the graph in the index.php page.