Changed include with include_once restayled list moving to typescript
Because we cannot include class more than one time, we used include_once
in the class for be sure to include the class only one time in the same page and
not throwing error.

Fixed list appearance

Because we decide to move to typescript we removed the javascript file.
Motive to move to typescript is that is a extension of javascript with many
good codestyle and feature (like async using promise).
added promise library and converted $ with jQuery for compatibility.