Fix error with no hosts to register
When we open target host window page, Eisen Frontend will try to connect to the Eisen Agent and
register ,the target host it found, in the Database.
The problem is if there is no Eisen Agent it will complain to not found any target host.
now it will just display the page without host and no error message will come out.

Added function for listing tasks.